requirements.txt
langchain_google_genai
langchain_openai
langchain_community
load-dotenv

